Move abstract base classes to dedicated base.py modules to prevent circular imports in the first place instead of working around them
