+++ thriftpy2/contrib/aio/transport/buffered.py
@@ -2,7 +2,7 @@
 import asyncio
 from io import BytesIO
 
-from . import TAsyncTransportBase
+from .base import TAsyncTransportBase
 
 
 class TAsyncBufferedTransport(TAsyncTransportBase):
